requirements.txt
django==4.2.11
flake8==6.1.0
python-docx==1.1.0
python-dotenv==1.0.1
